@@ -13,7 +13,7 @@
     "npm": "3.3.10"
   },
   "dependencies": {
-    "axios": "0.15.1",
+    "axios": "0.15.2",
     "babel-polyfill": "6.16.0",
     "bluebird": "3.4.6",
     "brace": "0.8.0",
@@ -29,10 +29,10 @@
     "js-cookie": "2.1.3",
     "json-loader": "0.5.4",
     "local-storage-fallback": "1.3.0",
-    "lodash": "4.16.4",
-    "material-ui": "0.16.0",
-    "moment": "2.15.1",
-    "moment-timezone": "0.5.6",
+    "lodash": "4.16.6",
+    "material-ui": "0.16.1",
+    "moment": "2.15.2",
+    "moment-timezone": "0.5.7",
     "mousetrap": "1.6.0",
     "normalize.css": "3.0.3",
     "numeral": "1.5.3",
@@ -40,25 +40,25 @@
     "promise-series": "1.0.0",
     "radium": "0.17.2",
     "raven-js": "3.7.0",
-    "react": "15.3.2",
-    "react-addons-create-fragment": "15.3.2",
-    "react-addons-linked-state-mixin": "15.3.2",
-    "react-addons-pure-render-mixin": "15.3.2",
-    "react-addons-test-utils": "15.3.2",
-    "react-addons-transition-group": "15.3.2",
-    "react-addons-update": "15.3.2",
+    "react": "15.4.0-rc.3",
+    "react-addons-create-fragment": "15.4.0-rc.3",
+    "react-addons-linked-state-mixin": "15.4.0-rc.3",
+    "react-addons-pure-render-mixin": "15.4.0-rc.3",
+    "react-addons-test-utils": "15.4.0-rc.3",
+    "react-addons-transition-group": "15.4.0-rc.3",
+    "react-addons-update": "15.4.0-rc.3",
     "react-copy-to-clipboard": "4.2.3",
-    "react-dom": "15.3.2",
-    "react-dropzone": "3.6.0",
+    "react-dom": "15.4.0-rc.3",
+    "react-dropzone": "3.7.2",
     "react-fittext": "0.1.1",
     "react-helmet": "3.1.0",
     "react-inlinesvg": "github:patrykkopycinski/react-inlinesvg#build",
     "react-input-mask": "0.7.5",
     "react-responsive-grid": "0.3.4",
-    "react-router": "3.0.0-beta.1",
+    "react-router": "3.0.0",
     "react-select": "0.9.1",
     "react-stickydiv": "3.4.19",
-    "react-tap-event-plugin": "1.0.0",
+    "react-tap-event-plugin": "github:Syncano/react-tap-event-plugin",
     "reflux": "0.3.0",
     "shortid": "2.2.6",
     "sort-object": "3.0.2",
@@ -70,22 +70,22 @@
   "devDependencies": {
     "autoprefixer-loader": "3.2.0",
     "aws-sdk": "2.4.5",
-    "babel-cli": "6.16.0",
-    "babel-core": "6.17.0",
-    "babel-eslint": "7.0.0",
-    "babel-loader": "6.2.5",
+    "babel-cli": "6.18.0",
+    "babel-core": "6.18.0",
+    "babel-eslint": "7.1.0",
+    "babel-loader": "6.2.7",
     "babel-plugin-add-module-exports": "0.2.1",
     "babel-plugin-lodash": "3.2.9",
     "babel-plugin-transform-react-constant-elements": "6.9.1",
     "babel-plugin-transform-react-inline-elements": "6.8.0",
-    "babel-preset-es2015": "6.16.0",
+    "babel-preset-es2015": "6.18.0",
     "babel-preset-react": "6.16.0",
     "babel-preset-react-hmre": "1.1.1",
     "babel-preset-stage-0": "6.16.0",
     "compass-mixins": "0.12.8",
     "css-loader": "0.25.0",
     "del": "1.2.1",
-    "eslint": "3.8.1",
+    "eslint": "3.9.1",
     "eslint-config-airbnb": "12.0.0",
     "eslint-friendly-formatter": "2.0.6",
     "eslint-loader": "1.6.0",
@@ -115,12 +115,12 @@
     "raw-loader": "0.5.1",
     "react-prism": "3.2.1",
     "sass-loader": "3.2.1",
-    "selenium-standalone": "5.7.2",
+    "selenium-standalone": "5.8.0",
     "semver": "5.3.0",
     "style-loader": "0.13.1",
     "through2": "2.0.1",
     "url-loader": "0.5.7",
-    "webpack": "1.13.2",
+    "webpack": "1.13.3",
     "webpack-dev-server": "1.16.2"
   },
   "scripts": {
@@ -128,6 +128,7 @@
     "build": "gulp build",
     "publish": "gulp publish",
     "clean": "gulp clean",
+    "reinstall": "rm -rf node_modules && npm install",
     "lint": "eslint src --quiet --ext .js --ext .jsx --fix",
     "lint-tests": "eslint test --quiet --ext .js --ext .jsx --fix",
     "lint-all": "eslint src test --quiet --ext .js --ext .jsx --fix",
