@@ -37,8 +37,8 @@ const HostingFilesView = React.createClass({
     const { instanceName } = this.props.params;
     const defaultHostingUrl = `https://${instanceName}.syncano.site/`;
     const hasDomains = hostingDetails && hostingDetails.domains.length > 0;
-    const customDomainUrl = hasDomains ? `https://${instanceName}--${hostingDetails.domains[0]}.syncano.site/` : null;
-    const hostingUrl = this.isDefaultHosting() ? defaultHostingUrl : customDomainUrl;
+    const customDomainUrl = hasDomains ? `https://${hostingDetails.domains[0]}--${instanceName}.syncano.site/` : null;
+    const hostingUrl = hostingDetails.is_default ? defaultHostingUrl : customDomainUrl;
 
     return hostingUrl;
   },
@@ -46,17 +46,7 @@ const HostingFilesView = React.createClass({
   getToolbarTitle() {
     const { hostingDetails, isLoading } = this.state;
 
-    return hostingDetails && !isLoading ? `Website Hosting: ${hostingDetails.label} (id: ${hostingDetails.id})` : '';
-  },
-
-  isDefaultHosting() {
-    const { hostingDetails } = this.state;
-
-    if (hostingDetails) {
-      return _.includes(hostingDetails.domains, 'default');
-    }
-
-    return false;
+    return hostingDetails && !isLoading ? `Website Hosting: ${hostingDetails.name} (id: ${hostingDetails.id})` : '';
   },
 
   handleBackClick() {
@@ -137,8 +127,13 @@ const HostingFilesView = React.createClass({
       currentFileIndex,
       isUploading,
       isDeleting,
-      errorResponses
+      errorResponses,
+      hostingDetails
     } = this.state;
+
+    if (!hostingDetails) {
+      return null;
+    }
 
     const hasFilesToUpload = filesToUpload.length > 0;
     const currentInstance = SessionStore.getInstance();
