@@ -50,8 +50,11 @@ export default Reflux.createStore({
   onCheckFolder(folder) {
     const { items } = this.data;
     const folderToCheck = _.find(items, { id: folder.id });
+    const isChecked = folder.checked;
 
-    folderToCheck.checked = !folder.checked;
+    _.forEach(folderToCheck.files, (file) => {
+      file.checked = !isChecked;
+    });
     this.trigger(this.data);
   },
 
@@ -61,7 +64,7 @@ export default Reflux.createStore({
     currentHostingId && Actions.fetchFiles(currentHostingId);
   },
 
-  setHostingId(hostingId) {
+  onSetHostingId(hostingId) {
     this.data.currentHostingId = hostingId;
   },
 
