@@ -38,7 +38,7 @@ const HostingList = React.createClass({
         title: 'Delete a Hosting',
         handleConfirm: HostingActions.removeHostings,
         items: getCheckedItems(),
-        itemLabelName: 'label',
+        itemLabelName: 'name',
         groupName: 'Hosting',
         isLoading
       }
