@@ -40,10 +40,24 @@ export default Reflux.createStore({
     });
   },
 
+  sortHostingDomains(domains, label, isDefault) {
+    // cname(s) -> default -> label
+    const cnameArray = _.without(domains, 'default', label);
+    const sortedDomains = [...cnameArray];
+
+    if (isDefault) {
+      sortedDomains.push('default');
+    }
+
+    sortedDomains.push(label);
+
+    return sortedDomains;
+  },
+
   setHosting(data) {
     const prepareHosting = (hosting) => {
-      hosting.isDefault = _.includes(hosting.domains, 'default');
-      hosting.cnameIndex = _.findIndex(hosting.domains, (domain) => domain !== 'default' && domain !== hosting.label);
+      hosting.domains = this.sortHostingDomains(hosting.domains, hosting.name, hosting.is_default);
+      hosting.cnameIndex = _.findIndex(hosting.domains, (domain) => domain !== 'default' && domain !== hosting.name);
       return hosting;
     };
 
