@@ -1,8 +1,7 @@
-import SyntheticUIEvent from 'react/lib/SyntheticUIEvent';
+import SyntheticUIEvent from 'react-dom/lib/SyntheticUIEvent';
 import _ from 'lodash';
 
 export default {
-
   isEvent(object) {
     return object instanceof SyntheticUIEvent;
   },
